+++ app/src/main/AndroidManifest.xml
@@ -12,14 +12,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".WaDirectActivity">
+        <activity android:name=".view.activity.WaDirectActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".CallLogActivity"></activity>
+        <activity android:name=".view.activity.CallLogActivity"></activity>
     </application>
 
 </manifest>
